# feat: Optimize link generation