--- a/app/api/schema/__init__.py
+++ b/app/api/schema/__init__.py
@@ -1,0 +1,10 @@
+# Monkey Patch Marshmallow JSONAPI
+from marshmallow_jsonapi.flask import Relationship
+
+
+def serialize(self, attr, obj, accessor=None):
+    if self.include_resource_linkage or self.include_data:
+        return super().serialize(attr, obj, accessor)
+    return self._serialize(None, attr, obj)
+
+Relationship.serialize = serialize