--- a/app/api/schema/__init__.py
+++ b/app/api/schema/__init__.py
@@ -4,7 +4,7 @@ from marshmallow_jsonapi.flask import Relationship
 
 def serialize(self, attr, obj, accessor=None):
     if self.include_resource_linkage or self.include_data:
-        return super().serialize(attr, obj, accessor)
+        return super(Relationship, self).serialize(attr, obj, accessor)
     return self._serialize(None, attr, obj)