Related: #6666 
Upstream: https://github.com/marshmallow-code/marshmallow-jsonapi/pull/278

Before:
/v1/events/ = 2.7 seconds == 2700 ms (20 events)
/v1/events/<id> = 250 ms

After:
/v1/events/ = 150 ms (20 events)
/v1/events/<id> = 30 ms